requirements.txt
flask==0.10.1
Pillow==2.9.0
sqlalchemy==1.0.8
